+++ src/Farmer/Builders/Builders.Functions.fs
@@ -13,6 +13,15 @@ open Farmer.Arm.KeyVault.Vaults
 open System
 
 type FunctionsRuntime = DotNet | DotNetIsolated | Node | Java | Python
+type DockerInfo = {
+    User: string
+    Password: SecureParameter
+    Url: Uri
+    StartupCommand: string
+}
+type PublishAs =
+    | Code
+    | DockerContainer of DockerInfo
 type FunctionsExtensionVersion = V1 | V2 | V3
 
 module private FunctionsConfig =
@@ -24,6 +33,7 @@ module private FunctionsConfig =
           Settings = state.Settings
           Cors = state.Cors
           Identity = state.Identity
+          KeyVaultReferenceIdentity = state.KeyVaultReferenceIdentity
           SecretStore = state.SecretStore
           ZipDeployPath = state.ZipDeployPath
           AlwaysOn = state.AlwaysOn
@@ -38,6 +48,7 @@ module private FunctionsConfig =
             Settings = config.Settings
             Cors = config.Cors
             Identity = config.Identity
+            KeyVaultReferenceIdentity = config.KeyVaultReferenceIdentity
             SecretStore = config.SecretStore
             ZipDeployPath = config.ZipDeployPath
             WorkerProcess = config.WorkerProcess }
@@ -54,8 +65,10 @@ type FunctionsConfig =
       Cors : Cors option
       StorageAccount : ResourceRef<FunctionsConfig>
       Runtime : FunctionsRuntime
+      PublishAs : PublishAs
       ExtensionVersion : FunctionsExtensionVersion
       Identity : ManagedIdentity
+      KeyVaultReferenceIdentity : UserAssignedIdentity Option
       SecretStore : SecretStore
       ZipDeployPath : string option
       AlwaysOn : bool
@@ -165,6 +178,15 @@ type FunctionsConfig =
                 if this.OperatingSystem = Windows then
                     "WEBSITE_CONTENTAZUREFILECONNECTIONSTRING", StorageAccount.getConnectionString this.StorageAccountName |> ArmExpression.Eval
                     "WEBSITE_CONTENTSHARE", this.Name.Value.ToLower()
+                match this.PublishAs with
+                | DockerContainer { User = us; Password = pass; Url = url } ->
+                    yield! [
+                        "DOCKER_REGISTRY_SERVER_URL", url.ToString()
+                        "DOCKER_REGISTRY_SERVER_USERNAME", us
+                        "DOCKER_REGISTRY_SERVER_PASSWORD", pass.ArmExpression.Eval()
+                    ]
+
+                | _ -> ()
               ]
               |> List.map Setting.AsLiteral
               |> List.append (
@@ -185,6 +207,7 @@ type FunctionsConfig =
               |> Map
 
               Identity = this.Identity
+              KeyVaultReferenceIdentity = this.KeyVaultReferenceIdentity
               Kind =
                 match this.OperatingSystem with
                 | Windows -> "functionapp"
@@ -235,7 +258,11 @@ type FunctionsConfig =
               PythonVersion = None
               Metadata = []
               ZipDeployPath = this.ZipDeployPath |> Option.map (fun x -> x, ZipDeploy.ZipDeployTarget.FunctionApp)
-              AppCommandLine = None
+              AppCommandLine =
+                match this.PublishAs with
+                | DockerContainer { StartupCommand = sc } ->
+                    Some sc
+                | _ -> None
               WorkerProcess = this.WorkerProcess }
             match this.ServicePlan with
             | DeployableResource this.Name resourceId ->
@@ -297,6 +324,8 @@ type FunctionsBuilder() =
           Settings = Map.empty
           Dependencies = Set.empty
           Identity = ManagedIdentity.Empty
+          KeyVaultReferenceIdentity = None
+          PublishAs = Code
           SecretStore = AppService
           Tags = Map.empty
           ZipDeployPath = None
@@ -316,6 +345,9 @@ type FunctionsBuilder() =
     /// Sets the runtime of the Functions host.
     [<CustomOperation "use_runtime">]
     member _.Runtime(state:FunctionsConfig, runtime) = { state with Runtime = runtime }
+    /// Sets the Publish as Code or Docker container information.
+    [<CustomOperation "publish_as">]
+    member _.PublishAs(state:FunctionsConfig, publishAs) = { state with PublishAs = publishAs }
     [<CustomOperation "use_extension_version">]
     member _.ExtensionVersion(state:FunctionsConfig, version) = { state with ExtensionVersion = version }
 
@@ -326,3 +358,8 @@ type FunctionsBuilder() =
         member _.Wrap state config = FunctionsConfig.FromCommon state config
 
 let functions = FunctionsBuilder()
+let docker (server: Uri) (user: string) (command: string): DockerInfo =
+    { User = user
+      Password = SecureParameter $"{user}-password"
+      Url = server
+      StartupCommand = command }