+++ src/Tests/Functions.fs
@@ -113,6 +113,17 @@ let tests = testList "Functions tests" [
         Expect.equal site.AppSettings.["FUNCTIONS_WORKER_RUNTIME"] (LiteralSetting "dotnet-isolated") "Should use dotnet-isolated functions runtime"
     }
 
+    test "Publish as docker container" {
+        let f = functions {
+            publish_as (DockerContainer (docker (new Uri("http://www.farmer.io")) "Robert Lewandowski" "do it")) }
+        let resources = (f :> IBuilder).BuildResources Location.WestEurope
+        let site = resources.[0] :?> Web.Site
+        Expect.equal site.AppSettings.["DOCKER_REGISTRY_SERVER_URL"] (LiteralSetting "http://www.farmer.io/") ""
+        Expect.equal site.AppSettings.["DOCKER_REGISTRY_SERVER_USERNAME"] (LiteralSetting "Robert Lewandowski") ""
+        Expect.equal site.AppSettings.["DOCKER_REGISTRY_SERVER_PASSWORD"] (LiteralSetting "[parameters('Robert Lewandowski-password')]") ""
+        Expect.equal site.AppCommandLine (Some "do it") ""
+    }
+
     test "Service plans support Elastic Premium functions" {
         let sp = servicePlan { name "test"; sku WebApp.Sku.EP2; max_elastic_workers 25 }
         let resources = (sp :> IBuilder).BuildResources Location.WestEurope